Parser without taking into account tools and errors
