--- .dist/data.csv
@@ -1,7 +1,0 @@
-N,I,R,U
-1,2,3,4
-2,3,4,5
-3,4,5,6
-4,5,6,7
-5,6,7,8
-6,7,8,9
